qno1 commit

+++ Source/Console/Classtest01/QuestionNo1/Program.cs
@@ -6,19 +6,19 @@ namespace QuestionNo1
     {
         static void Main(string[] args)
         {
-            float radius, width, height, side1, side2, side3;
+            double radius, width, height, side1, side2, side3;
             Console.WriteLine("Enter the radius for square");
-            radius = float.Parse(Console.ReadLine());
+            radius = double.Parse(Console.ReadLine());
             Console.WriteLine("Enter the width for rectangle");
-            width = float.Parse(Console.ReadLine());
+            width = double.Parse(Console.ReadLine());
             Console.WriteLine("Enter the height for rectangle");
-            height = float.Parse(Console.ReadLine());
+            height = double.Parse(Console.ReadLine());
             Console.WriteLine("Enter the side1 for triangle");
-            side1 = float.Parse(Console.ReadLine());
+            side1 = double.Parse(Console.ReadLine());
             Console.WriteLine("Enter the side2 for triangle");
-            side2 = float.Parse(Console.ReadLine());
+            side2 = double.Parse(Console.ReadLine());
             Console.WriteLine("Enter the side3 for triangle");
-            side3 = float.Parse(Console.ReadLine());
+            side3 = double.Parse(Console.ReadLine());
         }
     }
 }
